+++ src/components/structures/TopNav/TopNavItem.tsx
@@ -3,20 +3,25 @@ import { ComponentPropsWithoutRef } from "react";
 
 import styles from "./topNavItem.module.css";
 
-interface TopNavItemProps extends ComponentPropsWithoutRef<"div"> {
+export interface TopNavItemProps extends ComponentPropsWithoutRef<"div"> {
   hasSeparator?: boolean;
 }
 
 export default function TopNavItem({
   children,
+  className,
   hasSeparator,
   ...props
 }: TopNavItemProps) {
   return (
     <div
-      className={clsx(styles.root, {
-        [styles.hasSeparator]: hasSeparator,
-      })}
+      className={clsx(
+        styles.root,
+        {
+          [styles.hasSeparator]: hasSeparator,
+        },
+        className,
+      )}
       {...props}
     >
       {children}