+++ src/hooks/usePropRef.ts
@@ -1,17 +1,20 @@
-import { isFunction, isObject } from "lodash-es";
+import { isArray, isFunction, isObject } from "lodash-es";
 import { MutableRefObject, useCallback, useRef } from "react";
 
 export default function usePropRef<T = Node>(
-  propRef: React.ForwardedRef<T> | undefined,
+  propRef: React.ForwardedRef<T> | React.ForwardedRef<T>[] | undefined,
 ) {
   const innerRef: MutableRefObject<T | null> = useRef<T>(null);
   const setInnerRef = useCallback(
     (node: T) => {
-      if (propRef) {
-        if (isFunction(propRef)) {
-          propRef(node);
-        } else if (isObject(propRef)) {
-          propRef.current = node;
+      const refs = isArray(propRef) ? propRef : [propRef];
+      for (const ref of refs) {
+        if (ref) {
+          if (isFunction(ref)) {
+            ref(node);
+          } else if (isObject(ref)) {
+            ref.current = node;
+          }
         }
       }
       innerRef.current = node;