+++ src/wrappers/div.tsx
@@ -1,0 +1,12 @@
+import clsx from "clsx";
+
+import styles from "./div.module.css";
+import generateCssWrapper from "./generateCssWrapper.js";
+
+export const FlexColDiv = generateCssWrapper(
+  clsx(styles.flex, styles.flexColumn),
+);
+export const FlexAlignCenterDiv = generateCssWrapper(
+  clsx(styles.flex, styles.flexAlignCenter),
+);
+export const ContentMenuIcon = generateCssWrapper(styles.contentMenuIcon);